Set up micelle base class and test inheritance with spherical micelle class.
At this stage all variables and properties are stored as attributes
using the @property decorator. Only the variables (aggregation size,
temperature and number of carbon atoms in the tail) also have a setter
-- making the other properties not changeable.
This has the effect that volume, taillength and shape
specific properties are recalculated at each call.
This might actually not make sense here, because it slows down
the calculation and the benefit are not really needed in this
case? Will leave for now, as it's something new I learnt..