Add empirical interface free energy to base class.
The calculation of interfacial tension is not dependent on
shape and therefore done in the base class. Make area an
abstract attribute to allow for generic implementation of
interfacial free energy.
Add literature data for comparison and regression tests;
hard to get those to converge since extraction from pub.
data is imprecise, but plots confirm good results.