Implement boundaries in jacobian evaluation. Fix bug in jacobian
Jacobian was evaluated at same r_sph/cyl due to missing update and
wrong definitino in init -- fixed that.
Implemented a system in jacobian() to give back positive values when variables
left sensible spaces. This always returns a value so that the gradient
with the previous value sends back the optimiser: A more
positive value for jacobian value > 0 and more negative one else.