Vesicle (#3) Derivatives
* Uploadcommit start bilves derivatives

* Derivatives thickness inner - found bug

* Unfix bugfix and add thickness inner derivatives

* Finish property derivatives

* Finished derivatives and jacobian

Implementation tedious but equivalent to the one for rodlike

* Finish derivative based optimisation method for bil ves

It does not work as well as the objective function based one
which is obviously not good news. Differences in results
are nonexistent and it makes it less stable.

* Change bilayer vesicle standard optimisation method

Chang it back to objective because this is more stable
and is proven to give same radii up to pytest approx
specifications.